linker: use app_mem_regions header to populate defines
The linker is made generic for each SKU and its memory region
configurations are passed to the linker. This requires
a pre-processing state before passing it to the linker.

armclang linker does the preprocessing by itself. But for gcc,
it requires additional pre-processing state.

In cbuild env, it automatically does the pre-processing. But for Keil,
we need to add the pre-processing command in the Target build options.

Signed-off-by: Sudhir Sreedharan <sudhir@alifsemi.com>
Signed-off-by: Raj Ranjan <raj.ranjan@alifsemi.com>