HSCMP: Use P0_6 and P0_7 for HSCMP0 and HSCMP1
This commit includes the following changes:
1. The SEUART interface uses P0_0 for RX and P0_1 for TX.
Previously, HSCMP0 was configured to use P0_0 as its positive
input and HSCMP1 used P0_1.This pin conflict causes the SE
tool to become unresponsive during application flashing.
To avoid this issue, the pinmux configuration has been updated
in the application.
2.In RTE_device.h, the HSCMP0 positive input pin has been changed
to P0_6, and the HSCMP1 positive input pin to P0_7.

Signed-off-by: Nisarga AM <nisarga.am@alifsemi.com>