linker: add display buffer section in NS region for E1C
Added a display buffer section in the non-secure (NS) memory region
for the E1C platform. This addition includes a check for bulk SRAM
availability. If bulk SRAM is not available, the LCD frame buffer is
placed in the NS region instead of SRAM0, which is used by the Ensemble
platform when bulk SRAM is available.

Signed-off-by: Chandra Bhushan Singh <chandrabhushan.singh@alifsemi.com>